fix(spatial): validate arbitrary canvas plans
Keep area and Gaussian geometry unsigned through planning and use checked
signed deltas plus address-sized row offsets while sampling. Validate the
full RGBA descriptor before constructing plans and reserve Gaussian kernel
storage through a typed error path.

Build layout updates transactionally so failed descriptors or kernels retain
the prior plan. Cover signed dimension boundaries, large radii, offsets over
four GiB, host addressability, and failed-update sampling continuity.

+++ crates/hypercolor-core/src/spatial/plan.rs
@@ -49,11 +49,11 @@ pub struct PreparedBilinearSample {
 
 #[derive(Debug, Clone, Copy)]
 pub struct PreparedAreaSample {
-    pub center_x: i32,
-    pub center_y: i32,
-    pub radius: i32,
-    pub canvas_width: i32,
-    pub canvas_height: i32,
+    pub center_x: u32,
+    pub center_y: u32,
+    pub radius: u32,
+    pub canvas_width: u32,
+    pub canvas_height: u32,
     pub attenuation: u16,
 }
 
@@ -61,15 +61,15 @@ pub struct PreparedAreaSample {
 pub struct PreparedGaussianSamples {
     pub samples: Vec<PreparedGaussianSample>,
     pub weights: Vec<u16>,
-    pub weight_sum: u32,
+    pub weight_sum: u64,
 }
 
 #[derive(Debug, Clone, Copy)]
 pub struct PreparedGaussianSample {
-    pub center_x: i32,
-    pub center_y: i32,
-    pub radius: i32,
-    pub canvas_width: i32,
-    pub canvas_height: i32,
+    pub center_x: u32,
+    pub center_y: u32,
+    pub radius: u32,
+    pub canvas_width: u32,
+    pub canvas_height: u32,
     pub attenuation: u16,
 }
